Revert "Clear up the obsoleted cascading signal strength polling logic"
Some vendors appear to not use the standard UNSOL command to deliver
signal strength, or not send an UNSOL at all. In particular, on LG
devices with SDM845 or later chipset, LG's custom RIL stack does not
appear to be sending neither the AOSP nor the LG variant of
RIL_UNSOL_SIGNAL_STRENGTH whatsoever.

Because the polling logic was removed, these devices are currently
stuck on default SignalStrength values forever.

Bring it back.

This reverts commit 682dd4df959150fabdcb34efeef283e037bb2bc4.

Test: *#*#4636#*#* -> disable/enable radio power
Fixes: https://gitlab.com/LineageOS/issues/android/-/issues/6954
Change-Id: I149a129d39291fa67fa5c05c94fa7a4cb87e106b